requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bothub==0.1.30
bothub-cli==0.1.20
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
dateparser==1.0.0
dialogflow==1.1.0
docopt==0.6.2
elasticsearch==7.13.1
feedfinder2==0.0.4
feedparser==6.0.6
filelock==3.0.12
google-api-core==1.30.0
google-auth==1.31.0
google-news-crawler==0.3.9
googleapis-common-protos==1.53.0
GoogleNews==1.5.8
grpcio==1.38.0
idna==2.10
jieba3k==0.35.1
joblib==1.0.1
lxml==4.6.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.3
packaging==20.9
pandas==1.2.4
pathspec==0.8.1
Pillow==8.2.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.6
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
selenium==3.141.0
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
terminaltables==3.1.0
tinysegmenter==0.3
tldextract==3.1.0
tornado==6.1
tqdm==4.61.1
tzlocal==2.1
urllib3==1.26.5
wcwidth==0.2.5
zmq==0.0.0
